PLAT-68524: Add FullScreenPopup component
