+++ FullscreenPopup/FullscreenPopup.less
@@ -1,0 +1,14 @@
+// FullscreenPopup.less
+//
+@import "../styles/skin.less";
+
+.fullscreenPopup {
+	.inner {
+		height: 100%;
+		width: 100%;
+	}
+
+	.applySkins({
+		background-color: @agate-popup-bg-color;
+	});
+}
