+++ FullscreenPopup/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "FullscreenPopup.js"
+}
